Cast expanded timestamp to int after multiplication to prevent rounding errors associated with 7th decimal.

+++ tests/test_factory.py
@@ -71,8 +71,8 @@ class TestGet:
         timestamp = time.time()
         timestamp_dt = datetime.utcfromtimestamp(timestamp).replace(tzinfo=tz.tzutc())
 
-        assert self.factory.get(timestamp * 1e6) == timestamp_dt
-        assert self.factory.get(timestamp * 1e3) == timestamp_dt
+        assert self.factory.get(int(timestamp * 1e6)) == timestamp_dt
+        assert self.factory.get(int(timestamp * 1e3)) == timestamp_dt
 
     def test_one_arg_timestamp_with_tzinfo(self):
